requirements.txt
# Core Dependencies
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Cloud Storage
boto3==1.34.0
google-cloud-storage==2.10.0
s3fs==2023.12.2

# Data Warehouse
psycopg2-binary==2.9.9
google-cloud-bigquery==3.13.0
sqlalchemy==2.0.23
redshift-connector==2.0.915

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-amazon==8.11.0
apache-airflow-providers-google==10.10.1
apache-airflow-providers-postgres==5.7.1

# Data Processing
pyarrow==14.0.1
fastparquet==2023.10.1

# Configuration and Logging
pyyaml==6.0.1
python-dotenv==1.0.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
statsd==4.0.1

# Utilities
click==8.1.7
tqdm==4.66.1
schedule==1.2.0
